feat: support custom player & tidy up
