+++ src/components/default-player.ts
@@ -1,0 +1,6 @@
+export { default } from '@mux/mux-player-react';
+export * from '@mux/mux-player-react';
+export type {
+  MuxPlayerProps as DefaultPlayerProps,
+  MuxPlayerRefAttributes as DefaultPlayerRefAttributes,
+} from '@mux/mux-player-react';
